Stronger equality tests without casting trick.

+++ timemachine/fe/aligned_potential.py
@@ -369,7 +369,8 @@ class AlignedNonbondedInteractionGroup(AlignedPotential):
         w = interpolate_w_coord(src_w, dst_w, lamb).reshape(-1, 1)
         # (ytz): deal with spurious loss of precision when we interpolate nearly identical values:
         # eg. 0.3*x + 0.7*x ~= x
-        params = jnp.hstack([qse, w]).astype(np.float32).astype(np.float64)
+        params = jnp.hstack([qse, w])
+        # params = params.astype(np.float32).astype(np.float64)
         n_atoms = len(params)
         return NonbondedInteractionGroup(n_atoms, self.row_atom_idxs, self.beta, self.cutoff, self.col_atom_idxs).bind(
             params
